app/Http/Requests/Instructor/Courseディレクトリが肥大化しているための整理(yuki)
