+++ app/Http/Requests/Instructor/Course/DeleteRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Instructor;
+namespace App\Http\Requests\Instructor\Course;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class CourseDeleteRequest extends FormRequest
+class DeleteRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
