+++ app/Http/Requests/Instructor/Course/UpdateRequest.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\Http\Requests\Instructor\Course;
+
+use App\Rules\CourseStatusRule;
+use Illuminate\Foundation\Http\FormRequest;
+
+class UpdateRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            'course_id' => $this->route('course_id'),
+        ]);
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
+            'title' => ['required', 'string'],
+            'image' => ['mimes:jpg,png'],
+            'status' => ['required', 'string', new CourseStatusRule],
+        ];
+    }
+}
