--- app/Http/Requests/Instructor/CourseStoreRequest.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Instructor;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class CourseStoreRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'title' => ['required'],
-            'image' => ['required', 'file', 'image', 'mimes:jpeg,png,jpg', 'max:2048'],
-        ];
-    }
-}
